refactor: Controller 클래스 분리
- 요청에 대한 처리를 Controller가 수행하도록 설계
