+++ tomcat/src/main/java/org/apache/coyote/http11/RequestMapping.java
@@ -1,0 +1,20 @@
+package org.apache.coyote.http11;
+
+import org.apache.coyote.http11.controller.*;
+import org.apache.coyote.http11.request.HttpRequest;
+
+import java.util.List;
+
+public class RequestMapping {
+    private static List<Controller> CONTROLLER_INSTANCES = List.of(
+            new LoginGetController(), new LoginPostController(), new RegisterGetController(), new RegisterPostController(), new DefaultGetController()
+    );
+
+    public static Controller getController(HttpRequest request) {
+        return CONTROLLER_INSTANCES.stream()
+                .filter(controller -> controller.isSupported(request))
+                .findFirst()
+                .orElse(new DefaultController());
+    }
+}
+
